requirements.txt
django>=4.2
djangorestframework>=3.14
pandas>=2.0
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.0
drf-yasg>=1.21
django-widget-tweaks>=1.5
